Moves View's focus module
